key backup

+++ trixnity-core/src/commonMain/kotlin/net/folivo/trixnity/core/model/keys/RoomKeyBackupData.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class RoomKeyBackupData<T : RoomKeyBackupSessionData>(
     @SerialName("first_message_index")
-    val firstMessageIndex: Int,
+    val firstMessageIndex: Long,
     @SerialName("forwarded_count")
     val forwardedCount: Int,
     @SerialName("is_verified")
